Fix variables set to float datatype

+++ src/main/java/com/school/schedule/app/Teacher.java
@@ -5,7 +5,7 @@ public class Teacher extends Person{
     private Schedule schedule;
 
     // constructor
-    public Teacher(float id, String firstName, String lastName, Schedule schedule) {
+    public Teacher(int id, String firstName, String lastName, Schedule schedule) {
         super(id, firstName, lastName);
         this.schedule = schedule;
     }
